Since after swiping we have only positive numbers available it is possible now to use them for other purpose.
We can use existing array of elements as kind of Set that will indicate whenever given number was available.
Example: if we see number 3, we can make number in array under index 3 negative.

Then specification of the task, which states "smallest, positive, not existing number" ensures that the needle we are looking for will be smaller or equal to the length of array itself.

The +1 and -1 offsets are used to remove difference between 0-indexed arrays and the smallest positive number which is 1.

refs Day-4